Limited the tests that will be executed to the 'PyCOMTest's.
The pywin32 test facility is implemented as a hierarchy of test scripts
with the topmost script '.\pywin32_testall.py' calling among others the
next level script '.\com\win32com\test\testall.py' which finally calls
the PyComTest script '.\com\win32com\test\testPyComTest.py' among others.

This commit changes the files '.\pywin32_testall.py' and
'.\com\win32com\test\testall.py' so that they can be executed to run
only the PyComTest script '.\com\win32com\test\testPyComTest.py'.

The tests can either be executed by running the topmost script with:
(the topmost script internally sets the level to '1')

	python .\pywin32_testall.py -skip-adodbapi

or a level below specifying the level '1' explicitely with:

	python .\com\win32com\test\testall.py 1

It is also possible to directly invoking the PyCOMTest script with:

	python .\com\win32com\test\testPyComTest.py

which does not require the changes in this commit.

To successfully run the COM Record and Safearray tests, the PyComTest script
has to be modified to fix 2 failures unrelated to COM Records and Safearrays.
This is implemented in a seperate commit.

It is also required to compile and register the PyCOMTest server dll which
is described in the commit message of the above mentioned commit of the
failure fix for the PyComTest script.