Enable the test facility for COM Record and Safearray tests.
This commit fixes 2 failures of tests unrelated to COM Records and Safearrays
in the file '.\com\win32com\test\testPyComTest.py'.

With this fix it is possible to successfully execute the PyCOMTest script
directly, not using the upper level test scripts, with:

	python .\com\win32com\test\testPyComTest.py

To run the PyCOMTest script it is also required that the PyCOMTest server dll
was compiled and registered. The compilation can be accomplished by double
clicking the Visual Studio solution file

	'.\com\TestSources\PyCOMTest\PyComTest.sln'

and building the solution.

Registration or maybe re-registration requires an Administrator shell
and is done after the build with:

	python .\com\win32com\servers\test_pycomtest.py

For re-registration, if the dll was previously registered for a different
Python version, issue the following command before the registration:

	python .\com\win32com\servers\test_pycomtest.py --unregister